Rename pubmed -related modules
