+++ scripts/data/pmid_indra_counts_hasStatement_lite.csv
@@ -1,0 +1,101 @@
+pmid,pmcid,indra_evidences,indra_evidences_entity_constraint,indra_statements,indra_statements_entity_constraint
+33384007,PMC8091795,44,1,44,1
+33384043,PMC8259402,12,1,12,1
+33384255,,1,1,1,1
+33384330,PMC7812756,52,2,52,2
+33384332,PMC7812794,88,3,88,3
+33384338,PMC7826407,83,5,83,5
+33385212,PMC7780996,11,2,11,2
+33385332,PMC7854290,1,1,1,1
+33386047,,2,1,2,1
+33386711,PMC7776254,27,4,27,4
+33386716,PMC7776338,6,1,6,1
+33386742,,1,1,1,1
+33386784,,5,2,5,2
+33386801,,5,1,5,1
+33386802,PMC7990457,132,6,132,6
+33386804,PMC7990467,63,2,63,2
+33387066,,5,1,5,1
+33387069,,8,1,8,1
+33387440,PMC7786834,13,2,13,2
+33389929,PMC7944231,7,2,7,2
+33390025,,2,1,2,1
+33390034,,10,1,10,1
+33390050,,6,2,6,2
+33390051,PMC7808550,35,2,35,2
+33390052,PMC7856014,44,8,44,8
+33390067,,6,4,6,4
+33390366,PMC7902212,27,1,27,1
+33390389,,35,7,35,7
+33390425,,9,5,9,5
+33390474,,18,8,18,8
+33390476,,7,1,7,1
+33390479,,6,1,6,1
+33390480,,3,1,3,1
+33390482,,6,2,6,2
+33390490,PMC7946515,43,4,43,4
+33390519,,9,3,9,3
+33390525,,1,1,1,1
+33390526,,4,1,4,1
+33390536,,9,1,9,1
+33390540,,3,1,3,1
+33390545,,11,3,11,3
+33390552,,17,8,17,8
+33390554,,9,2,9,2
+33390585,PMC7791895,37,1,37,1
+33390631,PMC7729824,67,7,67,7
+33391389,PMC7774643,85,1,85,1
+33392353,PMC7701233,5,3,5,3
+33392482,PMC7773596,2,1,2,1
+33392504,PMC7764503,7,2,7,2
+33392508,PMC7764499,11,5,11,5
+33393288,,2,1,2,1
+33393325,,12,7,12,7
+33393435,,4,1,4,1
+33393496,PMC7773369,6,1,6,1
+33393497,PMC7773347,11,4,11,4
+33393510,PMC7773402,5,1,5,1
+33393537,,4,3,4,3
+33393560,PMC7937416,36,1,36,1
+33393570,,3,1,3,1
+33393580,,2,1,2,1
+33393764,,12,8,12,8
+33393765,,3,1,3,1
+33393767,PMC7909074,36,2,36,2
+33393781,,6,1,6,1
+33393784,,2,1,2,1
+33393815,,13,3,13,3
+33393837,PMC8254814,13,1,13,1
+33393839,PMC7801129,60,8,60,8
+33393852,PMC7801122,168,29,168,29
+33393859,,1,1,1,1
+33393955,,22,5,22,5
+33393968,PMC7797933,54,6,54,6
+33393970,PMC7794257,57,4,57,4
+33393974,PMC7797932,21,3,21,3
+33394045,PMC7916736,43,3,43,3
+33394582,PMC8315410,18,1,18,1
+33394604,PMC7905798,35,1,35,1
+33394628,PMC7959867,56,4,56,4
+33394671,PMC7945987,1,1,1,1
+33394672,PMC7789262,8,1,8,1
+33394687,PMC7861498,72,9,72,9
+33394707,PMC7895420,45,3,45,3
+33394716,,6,1,6,1
+33394718,,3,1,3,1
+33394730,,4,3,4,3
+33394732,PMC8221512,93,8,93,8
+33394733,PMC8240624,83,9,83,9
+33394740,PMC7880552,57,18,57,18
+33394748,PMC8480198,20,5,20,5
+33394752,PMC8018563,86,10,86,10
+33394753,PMC8049098,65,6,65,6
+33394970,,10,4,10,4
+33394974,,1,1,1,1
+33395067,PMC7952045,41,2,41,2
+33395136,,4,3,4,3
+33395186,,9,2,9,2
+33395272,PMC7844819,30,1,30,1
+33395281,,2,1,2,1
+33395287,PMC7816196,31,5,31,5
+33395289,PMC8023021,105,21,105,21
